\name{annotation.vandeVijver}

\docType{data}

\alias{annotation.vandeVijver}

\title{Some selected annotations for the breast cancer data from Rosetta Inpharmatics LLC and
Netherlands Cancer Institute, to serve as example data in the globaltest package.}

\description{A named list of 64 vectors of probe identifiers
appearing in the \code{\link{vandeVijver}}, linking the probes to
annotation terms from the Gene Ontology (GO) database. All 64 GO
terms are descendants of the GO term Cell Cycle (GO:0007049).}

\usage{data(annotation.vandeVijver)}

\format{A named list of vectors.}

\seealso{\code{\link{vandeVijver}}.}

\keyword{datasets}
