\name{contig2genbank}

\docType{data}

\alias{contig2genbank}

\title{Convert contig to Genbank accession number.}

\description{A named vector that can be used to convert the Contig
probe identifiers appearing in the \code{\link{vdVijver}} data into
Genbank accession numbers.}

\usage{data(contig2genbank)}

\format{A named vector.}

\seealso{\code{\link{vdVijver}}, \code{\link{RosettaAnnotation}}.}

\keyword{datasets}
