\name{exampleX}

\docType{data}

\alias{exampleX}

\title{Example expression data for package "globaltest".}

\description{A simulated gene expression dataset used in the 
examples of the package "globaltest".}

\usage{data(exampleX)}

\format{A matrix containing expressions for 1000 genes and 40
samples.}

\seealso{\code{\link{exampleY}}.}

\keyword{datasets}
