\name{exampleY}

\docType{data}

\alias{exampleY}

\title{Example clinical variable for package "globaltest".}

\description{A simulated clinical variable used used in the 
examples of the package "globaltest".}

\usage{data(exampleY)}

\format{A vector of length 40 containing values 0 and 1.}

\seealso{\code{\link{exampleX}}.}

\keyword{datasets}
