\name{gostructure.vandeVijver}

\docType{data}

\alias{gostructure.vandeVijver}

\title{A GOstructure object for the breast cancer data from Rosetta Inpharmatics LLC and
Netherlands Cancer Institute, to serve as example data in the
globaltest package.}

\description{A object with information on 64 annotation terms from
the Gene Ontology (GO) database, all descendants of the GO term Cell
Cycle (GO:0007049). The object contains information on ancestors,
descendants and associated genes.}

\usage{data(gostructure.vandeVijver)}

\format{A \code{\link{GOstructure}} object.}

\seealso{\code{\link{GOstructure}}, \code{\link{vandeVijver}}.}

\keyword{datasets}
