\name{Rosetta Annotation}

\docType{data}

\alias{RosettaAnnotation}

\title{Annotation for the breast cancer data from Rosetta Inpharmatics LLC and
Netherlands Cancer Institute.}

\description{A named list of vectors of probe identifiers appearing
in the \code{\link{vdVijver}}, linking the probes to annotation
terms.}

\usage{data(RosettaAnnotation)}

\format{A named list of vectors.}

\seealso{\code{\link{vdVijver}}, \code{\link{contig2genbank}}.}

\keyword{datasets}
