\name{annotation.vandeVijver}

\docType{data}

\alias{annotation.vandeVijver}

\title{Some selected annotations for the breast cancer data from Rosetta Inpharmatics LLC and
Netherlands Cancer Institute, to serve as example data in the globaltest package.}

\description{A named list of 15 vectors of probe identifiers appearing
in the \code{\link{vandeVijver}}, linking the probes to annotation
terms.}

\usage{data(RosettaAnnotation)}

\format{A named list of vectors.}

\seealso{\code{\link{vandeVijver}}.}

\keyword{datasets}
