\name{gt.result-class}
\docType{class}
\alias{p.value}
\alias{gt.result}
\alias{result}
\alias{gt.result-class}
\alias{show,gt.result-method}
\alias{[,gt.result-method}
\alias{length,gt.result-method}
\alias{hist,gt.result-method}
\alias{result,gt.result-method}
\alias{p.value,gt.result-method}
\alias{names,gt.result-method}
\alias{names<-,gt.result-method}
\alias{sort,gt.result-method}
\alias{combine,gt.result-method}
\alias{combine}
\alias{fit,gt.result-method}
\alias{fit}


\title{Class "gt.result" for results of the function globaltest}

\description{The class gt.result is the output of a call to
\code{\link{globaltest}} and the input of various plotting
functions to visualize the test result. }

\section{Slots}{
  \describe{
    \item{\code{res}:}{Object of class "matrix". Test results summary.}
    \item{\code{eX}:}{Object of class "matrix". The transformed data matrix.}
    \item{\code{genesets}:}{A list of vectors indicating the tested genes.}
    \item{\code{IminH}:}{Object of class "matrix". Needed when drawing the various
    diagnostic plots.}
    \item{\code{fit}:}{Object of class "list". This list contains one element which is of class "lm",
    "glm", "coxph" or "coxph.null". Thhis latter object contains the fitted model of the null hypothesis.}
    \item{\code{PermQs}:}{Object of class "matrix". Stores the permuted test
    statistics generated by a call to \code{\link{permutations}}.}
    \item{\code{SamplingPs}:}{Object of class "list". Stores the p values of random gene sets generated by
    a call to \code{\link{sampling}}.}
    }
}

\section{Methods}{
  \describe{
    \item{show}{(gt.result): Summarizes the test result.}
    \item{"["}{(gt.result): Extracts results of one or more
    pathways if multiple pathways were tested.}
    \item{length}{(gt.result): The number of pathways tested.}
    \item{p.value}{(gt.result): Extracts the p-values.}
    \item{sort}{(gt.result): Sorts the pathways to increasing p-values.}
    \item{result}{(gt.result): Extracts the results matrix.}
    \item{names}{(gt.result): Extracts the pathway names.}
    \item{names<-}{(gt.result): Changes the pathway names.}
    \item{combine}{(gt.result): Combines two {gt.result} objects to one,
    provided the data and model match.}
    \item{fit}{(gt.result): Extracts the fitted (adjust)model.}
    \item{hist}{(gt.result): Produces a histogram to visualize the
    permutations generated by \code{\link{permutations}}.}
    \item{geneplot}{(gt.result): The \code{\link{geneplot}} produces
    a plot to show the influence of individual
    genes on the test result produced by \code{\link{globaltest}}.}
    \item{sampleplot}{(gt.result): The \code{\link{sampleplot}}
    produces a plot to show the influence of individual
    samples on the test result produced by \code{\link{globaltest}}.}
    \item{permutations}{(gt.result): The function \code{\link{permutations}}
    recalculates the p-values using permutations of the outcome Y.}
    \item{sampling}{(gt.result): The function \code{\link{sampling}}
    compares the p-values with p-values of randomly
    generated pathways.}
    \item{checkerboard}{(gt.result): Produces a plot to visualize the test result
    produced by \code{\link{globaltest}} by showing the association
    between pairs of samples.}
    \item{regressionplot}{(gt.result): Produces a plot which can be used to visualize
    the effect of specific samples on the test result produced by
    \code{\link{globaltest}}.}
    }
}

\author{Jelle Goeman: \email{j.j.goeman@lumc.nl}; Jan Oosting}

\seealso{\code{\link{globaltest}}, \code{\link{sampleplot}},
\code{\link{geneplot}}, \code{\link{permutations}}.}

\keyword{methods}
